Move features instrumented tests to the features module
