+++ app/src/androidTest/java/com/maximo/douglas/githubconsumer/InstrumentedTestSuite.kt
@@ -1,8 +1,8 @@
 package com.maximo.douglas.githubconsumer
 
 import com.maximo.douglas.githubconsumer.activities.MainActivityTest
-import com.maximo.douglas.githubconsumer.features.gitpullrequests.SelectGitPullRequestFragmentTest
-import com.maximo.douglas.githubconsumer.features.gitrepositories.SelectGitRepositoryFragmentTest
+import com.maximo.douglas.features.gitpullrequests.SelectGitPullRequestFragmentTest
+import com.maximo.douglas.features.gitrepositories.SelectGitRepositoryFragmentTest
 import com.maximo.douglas.githubconsumer.navigation.NavigationTest
 import org.junit.runner.RunWith
 import org.junit.runners.Suite
