+++ commons/src/main/java/com/maximo/douglas/commons/testutils/ThreadUtil.kt
@@ -1,4 +1,4 @@
-package com.maximo.douglas.githubconsumer.testutils
+package com.maximo.douglas.commons.testutils
 
 object ThreadUtil {
