+++ commons/src/main/res/values/styles.xml
