+++ features/src/androidTest/java/com/maximo/douglas/features/gitpullrequests/SelectGitPullRequestFragmentTest.kt
@@ -1,4 +1,4 @@
-package com.maximo.douglas.githubconsumer.features.gitpullrequests
+package com.maximo.douglas.features.gitpullrequests
 
 import android.content.Intent
 import android.net.Uri
@@ -13,16 +13,16 @@ import androidx.test.espresso.intent.matcher.IntentMatchers
 import androidx.test.espresso.matcher.ViewMatchers
 import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
 import androidx.test.espresso.matcher.ViewMatchers.withText
-import br.com.argmax.githubconsumer.R
+import com.maximo.douglas.commons.testutils.FileUtils
+import com.maximo.douglas.commons.testutils.FileUtils.getJsonFromFile
+import com.maximo.douglas.commons.testutils.RecyclerViewMatcher.Companion.withRecyclerView
+import com.maximo.douglas.commons.testutils.ThreadUtil.waitViewToComplete
 import com.maximo.douglas.commons.utils.NavigationArgumentKeys.KEY_OWNER_LOGIN
 import com.maximo.douglas.commons.utils.NavigationArgumentKeys.KEY_REPOSITORY_NAME
 import com.maximo.douglas.commons.utils.StringUtils.compactStringWithDots
 import com.maximo.douglas.commons.utils.StringUtils.gitPullRequestClosedLabelStringFormat
 import com.maximo.douglas.commons.utils.StringUtils.gitPullRequestOpenLabelStringFormat
-import com.maximo.douglas.githubconsumer.testutils.FileUtils
-import com.maximo.douglas.githubconsumer.testutils.RecyclerViewMatcher.Companion.withRecyclerView
-import com.maximo.douglas.githubconsumer.testutils.ThreadUtil.waitViewToComplete
-import com.maximo.douglas.features.gitpullrequests.SelectGitPullRequestFragment
+import com.maximo.douglas.features.R
 import okhttp3.mockwebserver.Dispatcher
 import okhttp3.mockwebserver.MockResponse
 import okhttp3.mockwebserver.MockWebServer
@@ -172,7 +172,7 @@ class SelectGitPullRequestFragmentTest {
                 when (request.path) {
                     endpointToPullRequests -> return MockResponse()
                         .setResponseCode(200)
-                        .setBody(FileUtils.getJsonFromFile("jsonfiles/pullrequests/git_pull_request_api_response_cyc2018_cs_notes.json"))
+                        .setBody(getJsonFromFile("jsonfiles/pullrequests/git_pull_request_api_response_cyc2018_cs_notes.json"))
                 }
 
                 return MockResponse().setResponseCode(404)
