+++ features/src/androidTest/java/com/maximo/douglas/features/gitrepositories/SelectGitRepositoryFragmentTest.kt
@@ -1,16 +1,15 @@
-package com.maximo.douglas.githubconsumer.features.gitrepositories
+package com.maximo.douglas.features.gitrepositories
 
 import androidx.fragment.app.testing.FragmentScenario
 import androidx.lifecycle.Lifecycle
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.matcher.ViewMatchers.*
-import br.com.argmax.githubconsumer.R
+import com.maximo.douglas.commons.testutils.FileUtils.getJsonFromFile
+import com.maximo.douglas.commons.testutils.RecyclerViewMatcher.Companion.withRecyclerView
+import com.maximo.douglas.commons.testutils.ThreadUtil.waitViewToComplete
 import com.maximo.douglas.commons.utils.StringUtils
-import com.maximo.douglas.githubconsumer.testutils.FileUtils.getJsonFromFile
-import com.maximo.douglas.githubconsumer.testutils.RecyclerViewMatcher.Companion.withRecyclerView
-import com.maximo.douglas.githubconsumer.testutils.ThreadUtil.waitViewToComplete
-import com.maximo.douglas.features.gitrepositories.SelectGitRepositoryFragment
+import com.maximo.douglas.features.R
 import okhttp3.mockwebserver.MockResponse
 import okhttp3.mockwebserver.MockWebServer
 import org.junit.After
